Avatar Upload Form Improvements
The avatar image upload form was lacking validation and also it was impossible to tell if an image has been selected from the users device. Added the
bs-custom-file-input package. Added validation for the avatar image upload form in the UserAvatarController. Added JS code that checks if there is an error with the
avatar image upload field and shows the avatar modal if yes. The bs-custom-file-input shows the name of the file that is selected using the avatar image upload
form. The validation checks if the file is successfuly uploaded, is an image, is either jpeg, png, gif or webp and is under 2MB. The JS code that shows the avatar
modal improves the usability of the form.