Page Title & Breadcrumbs Updated In Dashboard
The page title and the breadcrumbs didn't have any functionality. Now the page title in the dashboard pages changes with every page. This was done using the
functionality of the blade templates. The breadcrumbs use a code that explodes the path of the URL and uses a foreach loop to generate the different breadcrumbs
elelemnts.