Added Projects View Page In Dashboard
Added projects view page in the dashboard. It shows a list of all the projects stored in the database. Also it uses pagination to reduce the size of the page. Other
things that were added are: A model for the projects table. A migration for the projects table. A factory for the projects model. A seeder for the projects table
using the factory for the projects model. Also vendor publish was called with the laravel-pagination tag. Also a route for the new projects page was added to the
web routes file.